HHH-9707 - Nulls first behavior on SQLServer database
