+++ hibernate-core/src/main/java/org/hibernate/dialect/SQLServer2008Dialect.java
@@ -25,6 +25,7 @@ package org.hibernate.dialect;
 
 import java.sql.Types;
 
+import org.hibernate.NullPrecedence;
 import org.hibernate.dialect.function.NoArgSQLFunction;
 import org.hibernate.type.StandardBasicTypes;
 
@@ -45,5 +46,27 @@ public class SQLServer2008Dialect extends SQLServer2005Dialect {
 		registerFunction(
 				"current_timestamp", new NoArgSQLFunction( "current_timestamp", StandardBasicTypes.TIMESTAMP, false )
 		);
+	}
+	
+	@Override
+	public String renderOrderByElement(String expression, String collation, String order, NullPrecedence nulls) {
+		final StringBuilder orderByElement = new StringBuilder();
+
+		if ( nulls != null && !NullPrecedence.NONE.equals( nulls ) ) {
+			// Workaround for NULLS FIRST / LAST support.
+			orderByElement.append( "case when " ).append( expression ).append( " is null then " );
+			if ( NullPrecedence.FIRST.equals( nulls ) ) {
+				orderByElement.append( "0 else 1" );
+			}
+			else {
+				orderByElement.append( "1 else 0" );
+			}
+			orderByElement.append( " end, " );
+		}
+
+		// Nulls precedence has already been handled so passing NONE value.
+		orderByElement.append( super.renderOrderByElement( expression, collation, order, NullPrecedence.NONE ) );
+
+		return orderByElement.toString();
 	}
 }
