+++ hibernate-core/src/main/java/org/hibernate/dialect/SQLServerDialect.java
@@ -27,7 +27,6 @@ import java.sql.Types;
 
 import org.hibernate.LockMode;
 import org.hibernate.LockOptions;
-import org.hibernate.NullPrecedence;
 import org.hibernate.dialect.function.AnsiTrimEmulationFunction;
 import org.hibernate.dialect.function.SQLFunctionTemplate;
 import org.hibernate.dialect.function.StandardSQLFunction;
@@ -209,30 +208,6 @@ public class SQLServerDialect extends AbstractTransactSQLDialect {
 	public int getInExpressionCountLimit() {
 		return PARAM_LIST_SIZE_LIMIT;
 	}
-	@Override
-	public String renderOrderByElement(String expression, String collation, String order, NullPrecedence nulls) {
-		final StringBuilder orderByElement = new StringBuilder();
-
-		if ( nulls != null && !NullPrecedence.NONE.equals( nulls ) ) {
-			// Workaround for NULLS FIRST / LAST support.
-			orderByElement.append( "case when " ).append( expression ).append( " is null then " );
-			if ( NullPrecedence.FIRST.equals( nulls ) ) {
-				orderByElement.append( "0 else 1" );
-			}
-			else {
-				orderByElement.append( "1 else 0" );
-			}
-			orderByElement.append( " end, " );
-		}
-
-		// Nulls precedence has already been handled so passing NONE value.
-		orderByElement.append( super.renderOrderByElement( expression, collation, order, NullPrecedence.NONE ) );
-
-		return orderByElement.toString();
-	}
 	
-	String renderOrderByElementDefaultBehavior(String expression, String collation, String order, NullPrecedence nulls) {
-		return super.renderOrderByElement( expression, collation, order, nulls );
-	}
 }
